move example player logic to gameobject
