+++ examples/src/main/java/dev/yeff/scenes/CollisionScene.java
@@ -1,18 +1,14 @@
 package dev.yeff.scenes;
 
-import dev.yeff.components.PlayerComponent;
 import dev.yeff.orbital.Game;
 import dev.yeff.orbital.ecs.GameObject;
-import dev.yeff.orbital.ecs.builders.GameObjectBuilder;
 import dev.yeff.orbital.ecs.components.collision.ColliderComponent;
 import dev.yeff.orbital.ecs.components.render.RenderShapeComponent;
 import dev.yeff.orbital.graphics.Color;
-import dev.yeff.orbital.graphics.Shapes;
 import dev.yeff.orbital.physics.collision.Collision;
 import dev.yeff.orbital.resources.ResourceManager;
 import dev.yeff.orbital.resources.Sprite;
 import dev.yeff.orbital.scenes.Scene;
-import org.joml.Vector2f;
 
 import java.util.Optional;
 
@@ -26,18 +22,18 @@ public class CollisionScene extends Scene {
     public void init(Game game) {
         sprite = ResourceManager.getSprite(getClass(), "assets/character_0000.png");
 
-        player = new GameObjectBuilder(this, "Player")
-                .withTransform(game.getScreenCenter(), new Vector2f(60.0f))
-                .withSprite(sprite)
-                .withCollider(Shapes.RECTANGLE)
-                .withComponent(new PlayerComponent())
-                .build();
-
-        obj2 = new GameObjectBuilder(this, "Object 1")
-                .withTransform(game.getScreenCenter(), new Vector2f(80.0f))
-                .withShape(Shapes.RECTANGLE, new Color(0, 255, 0))
-                .withCollider(Shapes.RECTANGLE)
-                .build();
+//        player = new GameObjectBuilder(this, "Player")
+//                .withTransform(game.getScreenCenter(), new Vector2f(60.0f))
+//                .withSprite(sprite)
+//                .withCollider(Shapes.RECTANGLE)
+//                .withComponent(new PlayerComponent())
+//                .build();
+//
+//        obj2 = new GameObjectBuilder(this, "Object 1")
+//                .withTransform(game.getScreenCenter(), new Vector2f(80.0f))
+//                .withShape(Shapes.RECTANGLE, new Color(0, 255, 0))
+//                .withCollider(Shapes.RECTANGLE)
+//                .build();
 
         addGameObject(game, obj2);
         addGameObject(game, player);
